Reverse out default-gateway change resetter
After all the other changes in the service from
debugging this, this seems to not be needed.

It always reconnects for me in about 20 seconds
without resetting the peers/paths

I left the resetter code in the history if we want to try to
get it down to a couple seconds someday. It does work
but you have to poll the default gateway code more
quickly and there may be other side effects.